fix: skip mirrored nodes at processing step
They weren't being processed before, but an empty list of children was
still being sent through. This resulted in extra empty lines in the
final output. Checks are done now to make sure that if there is only one
child node, that it is not a backlink before descending into the
children list.

There might be examples where there are multiple backlinks as children,
these might be processed mistakingly. This will have to wait to be fixed
until backlinks are properly handled.