fix(state): add newlines to ends of pages when testing
Newlines get added automatically at the end of each page (file) when
written to disk, but are absent in pages map, which means they are
missing for and fail the tests (since they compare against a file that
is read, and ends in newlines). So when testing, we need to add newlines
back to end of every page so they can pass.