refactor(md): renaming and commenting for clarity
Also, properties of node (n) are modified in place as they go along the
reformating and conversion assembly line. A lot is going on and it seems
actually more clear to just modify and move along to the next step
rather than try to keep track of multiple states of name and note
contents; you always see the state of things.

The contents of the node are trashed after this loop anyway so it's
already destructive.