fix(AppConfig): accidental closure bug when testing
Due to mocha or node caching something somewhere (I think), previous
instances of AppConfig were holding onto old values for the turndown
date and time custom rule internally. This prevented the toggling of
configuation settings on the fly just before a test using
`AppConfig.set` for both dateFormat and timeFormat.

This didn't actually matter in a one-off linear run of the script from
the command line, but it did mess up otherwise correct testing. Also, in
the long run, on-the-fly toggling of AppConfig needs to work correctly
for a webapp interface where settings are toggled continuously with the
same instance.